travelplanner/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
bleach==6.2.0
certifi==2024.12.14
distro==1.9.0
Django
django-filter==24.3
djangorestframework==3.15.2
drf-yasg==1.21.8
exceptiongroup==1.2.2
groq==0.15.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pydantic==2.10.5
pydantic_core==2.27.2
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
webencodings==0.5.1
